Add files via upload
Program Description

    Write a program that prints out your name, the course ID of this class, what programming/computer courses you've taken.
    Ask the user for two numbers.
    Load the user numbers into variables.
    Show the sum of the two numbers, the difference, the product and the quotient.
        For the difference subtract the second number from the first, for the quotient, use the first number as the numerator(dividend) and the second as the denominator(divisor).

 

Sample Output:

My name is Glenn, I'm in course 123456. I've taken:

C Programming

C++ Programming

 

Give me a number: 7

Give me a second number: 2

 

The sum is 9.

The difference is 5.

The product is 14.

The quotient is 3.

 
Homework Submission

You need to turn your code in to get credit. Your code file should be named with the following convention, Homework0YourName.c, for example Homework0GlennCuevas.c if I were turning it in.

If your code does not compile, you need to note than in the submission notes so that I don't try to compile it.