Add files via upload
Calculates the amount earned pre and post tax for 5 employees. The purpose of this project was to learn the storing of C-strings, floats, and if statements. 

If I were to redo this assignment, I would change the printf statements from line 161 to line 194 to print variables instead of text since the memory allocation for the repeated text outputs will be larger than the reused variable. I'm unsure as to which would be faster though.